+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 * Switched to NNUE-based evaluation with a (768->256)x2->1 net (223.69 +- 33.98)
+* Use `align(64)` for network weights (5.51 +- 4.06)
 * Encode double pushes in `Move` instead of checking in `make_move` (7.46 +- 6.47 in regression)
 
 ### Misc